edge_controller/pyproject.toml
[tool.poetry]
name = "edge-controller"
version = "0.1.0"
description = ""
authors = ["mbw <matthew@mh-white.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
paho-mqtt = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
